cir_system/requirements.txt
gradio>=3.50.0
pypdf>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pillow>=10.0.0
requests>=2.28.0
langchain>=0.1.0
langchain-core>=0.1.0
ollama>=0.1.0
openpyxl>=3.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
